# server/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the Swagger Documentation of SafeWorks server API",
    "version": "1.0.0",
    "title": "SafeWorks API"
  },
  "host": "localhost:3001",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/jobs": {
      "post": {
        "tags": [
          "JobAd"
        ],
        "description": "Endpoint to add a new JobAd",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "jobAd",
            "in": "body",
            "description": "JobAd data to add",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddJobAd"
            }
          },
          {
            "name": "files",
            "in": "formData",
            "type": "file",
            "description": "Images file to upload with the JobAd data",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created JobAd UID",
            "schema": {
              "$ref": "#/definitions/CreatedJobAd"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "JobAd"
        ],
        "description": "Endpoint to update a JobAd",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "job",
            "in": "body",
            "description": "JobAd data to update",
            "required": true,
            "schema": {
              "$ref": "#/definitions/AddJobAd"
            }
          },
          {
            "name": "files",
            "in": "formData",
            "type": "file",
            "description": "Images file to upload with the JobAd data",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Updated JobAd UID",
            "schema": {
              "$ref": "#/definitions/CreatedJobAd"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "JobAd"
        ],
        "description": "Endpoint to get all JobAds",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "type": "int"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": true,
            "type": "int"
          }
        ],
        "responses": {
          "200": {
            "description": "All found JobAds",
            "schema": {
              "$ref": "#/definitions/AllJobAds"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "JobAd"
        ],
        "description": "Endpoint to delete a JobAd",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "description": "JobAd UID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/jobs/{uid}": {
      "get": {
        "tags": [
          "JobAd"
        ],
        "description": "Endpoint to get a JobAd by ID",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "JobAd UID"
          }
        ],
        "responses": {
          "200": {
            "description": "JobAd response object",
            "schema": {
              "$ref": "#/definitions/JobAd"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/jobs/worker/{uid}": {
      "get": {
        "tags": [
          "JobAd"
        ],
        "description": "Endpoint to get all JobAds from a Worker",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Worker User UID"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number",
            "required": true,
            "type": "int"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "required": true,
            "type": "int"
          }
        ],
        "responses": {
          "200": {
            "description": "All found JobAds",
            "schema": {
              "$ref": "#/definitions/AllJobAds"
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/jobs/find/{term}": {
      "get": {
        "tags": [
          "JobAd"
        ],
        "description": "Endpoint to get find JobAds by term",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "term",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Term to search"
          }
        ],
        "responses": {
          "200": {
            "description": "All found JobAds",
            "schema": {
              "$ref": "#/definitions/AllJobAds"
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "JobAd": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 200
        },
        "job": {
          "type": "object",
          "properties": {
            "uid": {
              "type": "string",
              "example": "1YS5YRafCyUThV1kEucM"
            },
            "worker": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Lucas"
                },
                "id": {
                  "type": "string",
                  "example": "NESmmmshXaSSeYHMB5FtU0l2hTg2"
                }
              }
            },
            "title": {
              "type": "string",
              "example": "Teste"
            },
            "description": {
              "type": "string",
              "example": "Descrição"
            },
            "category": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Carpinteiro"
                },
                "id": {
                  "type": "number",
                  "example": 1
                }
              }
            },
            "district": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Água Verde"
                },
                "id": {
                  "type": "number",
                  "example": 1
                }
              }
            },
            "price": {
              "type": "number",
              "example": 10
            },
            "price_type": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Dia"
                },
                "id": {
                  "type": "number",
                  "example": 1
                }
              }
            },
            "displacement_fee": {
              "type": "number",
              "example": 0
            },
            "delivery_time": {
              "type": "number",
              "example": 1
            },
            "media": {
              "type": "array",
              "example": [
                "https://media1.com"
              ],
              "items": {
                "type": "string"
              }
            },
            "contracts": {
              "type": "array",
              "example": [
                "23S5YRafCyUThV1kEu23"
              ],
              "items": {
                "type": "string"
              }
            },
            "expired": {
              "type": "boolean",
              "example": false
            },
            "created": {
              "type": "string",
              "example": "06/07/2023 18:32:28"
            },
            "modified": {
              "type": "string",
              "example": "07/07/2023 18:32:28"
            },
            "deleted": {}
          }
        }
      }
    },
    "AllJobAds": {
      "type": "object",
      "properties": {
        "jobs": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "uid": {
                "type": "string",
                "example": "1YS5YRafCyUThV1kEucM"
              },
              "worker": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Lucas"
                  },
                  "id": {
                    "type": "string",
                    "example": "NESmmmshXaSSeYHMB5FtU0l2hTg2"
                  }
                }
              },
              "title": {
                "type": "string",
                "example": "Teste"
              },
              "description": {
                "type": "string",
                "example": "Descrição"
              },
              "category": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Carpinteiro"
                  },
                  "id": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "district": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Água Verde"
                  },
                  "id": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "price": {
                "type": "number",
                "example": 10
              },
              "price_type": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Dia"
                  },
                  "id": {
                    "type": "number",
                    "example": 1
                  }
                }
              },
              "displacement_fee": {
                "type": "number",
                "example": 0
              },
              "delivery_time": {
                "type": "number",
                "example": 1
              },
              "media": {
                "type": "array",
                "example": [
                  "https://media1.com"
                ],
                "items": {
                  "type": "string"
                }
              },
              "contracts": {
                "type": "array",
                "example": [
                  "23S5YRafCyUThV1kEu23"
                ],
                "items": {
                  "type": "string"
                }
              },
              "expired": {
                "type": "boolean",
                "example": false
              },
              "created": {
                "type": "string",
                "example": "06/07/2023 18:32:28"
              },
              "modified": {
                "type": "string",
                "example": "07/07/2023 18:32:28"
              },
              "deleted": {}
            }
          }
        },
        "total": {
          "type": "number",
          "example": 2
        },
        "currentPage": {
          "type": "number",
          "example": 1
        }
      }
    },
    "AddJobAd": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string",
          "example": "Teste"
        },
        "description": {
          "type": "string",
          "example": "Descrição"
        },
        "category": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 6
            },
            "name": {
              "type": "string",
              "example": "Carpinteiro"
            }
          }
        },
        "district": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 1
            },
            "name": {
              "type": "string",
              "example": "Água Verde"
            }
          }
        },
        "price": {
          "type": "number",
          "example": 11
        },
        "price_type": {
          "type": "object",
          "properties": {
            "id": {
              "type": "number",
              "example": 1
            },
            "name": {
              "type": "string",
              "example": "Dia"
            }
          }
        },
        "worker": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string",
              "example": "NESmmmshXaSSeYHMB5FtU0l2hTg2"
            },
            "name": {
              "type": "string",
              "example": "Lucas"
            }
          }
        }
      },
      "required": [
        "title",
        "description",
        "category",
        "district",
        "price",
        "price_type",
        "worker"
      ]
    },
    "CreatedJobAd": {
      "type": "object",
      "properties": {
        "statusCode": {
          "type": "number",
          "example": 201
        },
        "jobAd": {
          "type": "string",
          "example": "1YS5YRafCyUThV1kEucM"
        }
      }
    },
...
  }
}